endpoint/requirements.txt
python-dotenv>=1.0.0
windows-use>=0.1.0
